app/requirements.txt
attrs==21.2.0
falcon==1.4.1
falcon-autocrud==1.0.37
greenlet==1.1.0
jsonschema==3.2.0
psycopg2-binary==2.9.1
pyrsistent==0.18.0
python-mimeparse==1.6.0
six==1.16.0
SQLAlchemy==1.4.20
uWSGI==2.0.19.1
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.0
falcon==1.4.1
falcon-autocrud==1.0.37
greenlet==1.1.0
gunicorn==20.1.0
idna==3.2
jsonschema==3.2.0
psycopg2-binary==2.9.1
pyrsistent==0.18.0
python-mimeparse==1.6.0
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.20
urllib3==1.26.6
alembic==1.6.5
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.0
falcon==1.4.1
falcon-autocrud==1.0.37
greenlet==1.1.0
idna==3.2
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.18.0
python-editor==1.0.4
python-mimeparse==1.6.0
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.20
urllib3==1.26.6
uWSGI==2.0.19.1
